fragment with layout files were created for: settings screen, search screen, main activity was rewritten for using fragments
