+++ app/build.gradle
@@ -69,6 +69,7 @@ dependencies {
     // Navigation libraries
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+    implementation "androidx.fragment:fragment-ktx:1.5.6"
 
     implementation "androidx.datastore:datastore-preferences:1.0.0"
