+++ app/src/main/java/bes/max/playlistmaker/presentation/mediateka/fragments/FavoriteTracksFragment.kt
@@ -1,10 +1,7 @@
 package bes.max.playlistmaker.presentation.mediateka.fragments
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import bes.max.playlistmaker.R
 import bes.max.playlistmaker.databinding.FragmentFavoriteTracksBinding
 import bes.max.playlistmaker.presentation.mediateka.viewmodels.FavoriteTracksViewModel
 import bes.max.playlistmaker.presentation.utils.BindingFragment
@@ -19,13 +16,6 @@ class FavoriteTracksFragment : BindingFragment<FragmentFavoriteTracksBinding>() 
         container: ViewGroup?
     ): FragmentFavoriteTracksBinding {
         return FragmentFavoriteTracksBinding.inflate(inflater, container, false)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_favorite_tracks, container, false)
     }
 
     companion object {
