+++ app/src/main/java/bes/max/playlistmaker/presentation/mediateka/fragments/PlaylistFragment.kt
@@ -1,10 +1,7 @@
 package bes.max.playlistmaker.presentation.mediateka.fragments
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import bes.max.playlistmaker.R
 import bes.max.playlistmaker.databinding.FragmentPlaylistBinding
 import bes.max.playlistmaker.presentation.mediateka.viewmodels.PlaylistViewModel
 import bes.max.playlistmaker.presentation.utils.BindingFragment
@@ -19,13 +16,6 @@ class PlaylistFragment : BindingFragment<FragmentPlaylistBinding>() {
         container: ViewGroup?
     ): FragmentPlaylistBinding {
         return FragmentPlaylistBinding.inflate(inflater, container, false)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_playlist, container, false)
     }
 
     companion object {
