+++ app/src/main/res/values-night/styles.xml
@@ -1,26 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="main_activity_buttons" parent="Widget.MaterialComponents.Button">
-        <item name="android:layout_width">match_parent</item>
-        <item name="android:layout_height">0dp</item>
-        <item name="android:layout_weight">1</item>
-        <item name="android:backgroundTint">@color/white</item>
-        <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">22sp</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textAllCaps">false</item>
-        <item name="cornerRadius">24dp</item>
-        <item name="iconPadding">8dp</item>
-        <item name="iconTint">@color/black</item>
-        <item name="iconGravity">textStart</item>
-    </style>
-
     <style name="screen_title_textview">
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="android:textColor">@color/white</item>
         <item name="android:textSize">22sp</item>
-        <item name="android:layout_marginStart">24dp</item>
+        <item name="android:layout_marginStart">16dp</item>
     </style>
 
     <style name="toolbar_back_arrow">
