+++ app/src/main/res/values/strings.xml
@@ -39,7 +39,8 @@
 
     <string name="fragment_mediateka_tab_title_favorite_tracks">Избранные треки</string>
     <string name="fragment_mediateka_tab_title_playlist">Плейлисты</string>
-
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 
 </resources>
