+++ app/src/main/res/values/styles.xml
@@ -1,21 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="main_activity_buttons" parent="Widget.MaterialComponents.Button.UnelevatedButton">
-        <item name="android:layout_width">match_parent</item>
-        <item name="android:layout_height">0dp</item>
-        <item name="android:layout_weight">1</item>
-        <item name="android:backgroundTint">@color/white</item>
-        <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">22sp</item>
-        <item name="android:fontFamily">@font/ys_display_medium</item>
-        <item name="android:textAllCaps">false</item>
-        <item name="cornerRadius">24dp</item>
-        <item name="iconPadding">8dp</item>
-        <item name="iconTint">@color/icons_main_screen</item>
-        <item name="iconGravity">textStart</item>
-    </style>
-
     <style name="toolbar_back_arrow">
         <item name="android:layout_width">@dimen/back_arrow_size</item>
         <item name="android:layout_height">@dimen/back_arrow_size</item>
@@ -27,7 +12,7 @@
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">22sp</item>
-        <item name="android:layout_marginStart">24dp</item>
+        <item name="android:layout_marginStart">16dp</item>
     </style>
 
     <style name="settings_activity_image_view">
